[Test] Move capability declaration tests to common test suite
Move backend capability declaration coverage out of the OpenReg-only suite so static and mocked backend declarations are exercised by the common testing suite. Keep OpenReg-specific gating coverage in place and align missing-capability expectations with SkipTest behavior.

Test Plan:

```text
.venv/bin/python -m py_compile test/cpp_extensions/open_registration_extension/torch_openreg/tests/test_testing.py test/test_testing.py torch/testing/_internal/common_device_type.py
.venv/bin/python test/test_testing.py TestFP64CapabilityDeclarations -v
CPATH=/home/daniel/workspace/pytorch-dev/worktrees/pytorch/pr191919-main-replay-20260828/third_party/pybind11/include .venv/bin/python test/run_test.py --openreg --verbose
XDG_DATA_HOME=/home/daniel/workspace/pytorch-dev/worktrees/pytorch/pr191919-main-replay-20260828/agent_space/xdg-data UV_CACHE_DIR=/home/daniel/workspace/pytorch-dev/worktrees/pytorch/pr191919-main-replay-20260828/agent_space/uv-cache UV_TOOL_DIR=/home/daniel/workspace/pytorch-dev/worktrees/pytorch/pr191919-main-replay-20260828/agent_space/uv-tools UV_TOOL_BIN_DIR=/home/daniel/workspace/pytorch-dev/worktrees/pytorch/pr191919-main-replay-20260828/agent_space/uv-bin UV_PYTHON_INSTALL_DIR=/home/daniel/workspace/pytorch-dev/worktrees/pytorch/pr191919-main-replay-20260828/agent_space/uv-python PATH=/home/daniel/workspace/pytorch-dev/worktrees/pytorch/pr191919-main-replay-20260828/.venv/bin:$PATH .venv/bin/spin quicklint -- test/cpp_extensions/open_registration_extension/torch_openreg/tests/test_testing.py test/test_testing.py torch/testing/_internal/common_device_type.py
git diff --check
```

Authored with assistance from Codex, an AI coding assistant.
